Bug fixes for links
Change updates the link create process. The change adds a couple of
guards to the link process in both the wheel_builder.py and
html_indexer.py

+++ yaprt/wheel_builder.py
@@ -803,11 +803,12 @@ class WheelBuilder(utils.RepoBaseClass):
                         raise exp
 
             if not os.path.islink(link_path):
-                # Create the symlink
-                os.symlink(
-                    os.path.relpath(full_wheel_path),
-                    os.path.join(self.args['link_dir'], wheel_name)
-                )
+                if os.path.isfile(full_wheel_path):
+                    # Create the symlink
+                    os.symlink(
+                        os.path.relpath(full_wheel_path),
+                        os.path.join(self.args['link_dir'], wheel_name)
+                    )
 
     def _package_clean(self, package, files):
         """Remove links for a given package name if found.
